+++ lib/AudioLeds/effects/RainbowEffect.h
@@ -10,7 +10,7 @@ class RainbowEffect: public AbstractEffect
     private:
         uint8_t hue;
     public:
-        virtual void loop(uint16_t fill_size) override;
+        virtual void loop(uint8_t fill_size) override;
 };
 
 #endif
